[FD-388] feat: 아키텍쳐 변경으로 인한 테스트 코드 수정
