+++ back-end/src/test/java/com/feedhanjum/back_end/schedule/controller/ScheduleControllerTest.java
@@ -4,6 +4,7 @@ import com.feedhanjum.back_end.schedule.controller.dto.ScheduleRequest;
 import com.feedhanjum.back_end.schedule.domain.Todo;
 import com.feedhanjum.back_end.schedule.service.ScheduleService;
 import com.feedhanjum.back_end.schedule.service.dto.ScheduleRequestDto;
+import com.feedhanjum.back_end.teamplanorchestration.service.TeamPlanOrchestrationService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -28,6 +29,9 @@ public class ScheduleControllerTest {
     private ScheduleController scheduleController;
 
     @Mock
+    private TeamPlanOrchestrationService teamPlanOrchestrationService;
+
+    @Mock
     private ScheduleService scheduleService;
     @Test
     @DisplayName("일정 생성 컨트롤러 성공 테스트")
@@ -38,7 +42,7 @@ public class ScheduleControllerTest {
         Todo hehe = new Todo("hehe");
         ScheduleRequest request = new ScheduleRequest("haha", LocalDateTime.now(), LocalDateTime.now().plusDays(10), List.of(hehe));
         ScheduleRequestDto scheduleRequestDto = new ScheduleRequestDto(request);
-        doNothing().when(scheduleService).createSchedule(memberId, teamId, scheduleRequestDto);
+        doNothing().when(teamPlanOrchestrationService).createSchedule(memberId, teamId, scheduleRequestDto);
 
         // when
         ResponseEntity<Void> result = scheduleController.createSchedule(memberId, teamId, request);
@@ -46,7 +50,7 @@ public class ScheduleControllerTest {
         // then
         assertThat(result).isNotNull();
         assertThat(result.getStatusCode()).isEqualTo(HttpStatus.CREATED);
-        verify(scheduleService).createSchedule(memberId, teamId, new ScheduleRequestDto(request));
+        verify(teamPlanOrchestrationService).createSchedule(memberId, teamId, new ScheduleRequestDto(request));
     }
     @Test
     @DisplayName("스케줄 수정 요청 성공")
